fix: don't erase xrefs that we need (#482)
close #481 

Our QC process includes removal of ALL values in a field if there are
more than a certain number, due to old dyanmodb memory restrictions that
we should probably update.

In this case, this record had a ton of product codes added, which was
causing us to dump the xrefs that we need for grouping (ie to rxnorm)
along with the ones we didn't. This PR updates the QC procedure to first
try to prune xrefs down to the essential ones.

Unfortunately we also have to drop all the hemonc-provided trade names
for imatinib. This would be fixed if we could take time to reinvestigate
size limitations on dynamodb items (i'm pretty sure they're much less
strict now) or switch to a different DB.